requirements.txt
# Core ML
tensorflow>=2.12
numpy

# Data prep
pyyaml
pillow

# Metrics
scikit-learn

# Realtime / conveyor (webcam, industrial camera capture)
opencv-python>=4.8
